# task-service/internal/interfaces/http/gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: Persists Task execution records and streams live updates.
  title: KAPE Task Service
  version: 1.0.0
servers:
- url: http://localhost:8080
tags:
- name: tasks
- name: handlers
paths:
  /tasks:
    get:
      operationId: listTasks
      parameters:
      - explode: true
        in: query
        name: handler
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/TaskStatus"
        style: form
      - explode: true
        in: query
        name: since
        required: false
        schema:
          format: date-time
          type: string
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          default: received_at:desc
          enum:
          - received_at:asc
          - received_at:desc
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
          description: Task list
      summary: List tasks with optional filters
      tags:
      - tasks
    post:
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
          description: Task created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request body
      summary: Create a task record
      tags:
      - tasks
  /tasks/stream:
    get:
      operationId: streamTasks
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: Server-Sent Events stream; each event is a Task JSON object
      summary: SSE stream of live task updates
      tags:
      - tasks
  /tasks/decisions:
    get:
      operationId: getDecisions
      parameters:
      - explode: true
        in: query
        name: handler
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: since
        required: true
        schema:
          format: date-time
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionDistribution"
          description: Decision distribution
      summary: Decision distribution per handler
      tags:
      - tasks
  /tasks/bulk/status:
    patch:
      operationId: bulkUpdateStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateStatusRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpdateStatusResponse"
          description: Affected task IDs
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request
      summary: Bulk update status for operator-selected tasks
      tags:
      - tasks
  /tasks/{id}:
    delete:
      operationId: deleteTask
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: Discard stale event — deletes the task record with no terminal state
      tags:
      - tasks
    get:
      operationId: getTask
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
          description: Task
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: Fetch a single task
      tags:
      - tasks
  /tasks/{id}/status:
    patch:
      operationId: updateTaskStatus
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
          description: Updated task
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid status transition
      summary: Update task status
      tags:
      - tasks
  /tasks/{id}/retry:
    post:
      operationId: retryTask
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "501":
          description: Not implemented until Phase 7
      summary: Retry a task (Phase 7 — stub returns 501)
      tags:
      - tasks
  /tasks/{id}/lineage:
    get:
      operationId: getTaskLineage
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Task"
                type: array
          description: Lineage chain ordered root to latest
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: Full retry lineage chain for a task
      tags:
      - tasks
  /handlers:
    get:
      operationId: listHandlers
      responses:
        "501":
          description: Not implemented — handler aggregates are owned by Prometheus
      summary: Per-handler aggregates (stub — Prometheus integration pending per 0008
        spec)
      tags:
      - handlers
components:
  schemas:
    TaskStatus:
      enum:
      - Processing
      - Completed
      - Failed
      - SchemaValidationFailed
      - ActionError
      - UnprocessableEvent
      - PendingApproval
      - Timeout
      - Retried
      type: string
    ActionResult:
      example:
        name: name
        dry_run: true
        type: type
        error: error
        status: status
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
        dry_run:
          type: boolean
        error:
          nullable: true
          type: string
      required:
      - dry_run
      - name
      - status
      - type
      type: object
    TaskError:
      example:
        schema: schema
        raw: raw
        detail: detail
        type: type
        traceback: traceback
      properties:
        type:
          type: string
        detail:
          type: string
        schema:
          nullable: true
          type: string
        raw:
          nullable: true
          type: string
        traceback:
          nullable: true
          type: string
      required:
      - detail
      - type
      type: object
    Task:
      example:
        cluster: cluster
        handler: handler
        retry_of: retry_of
        event_raw:
          key: ""
        otel_trace_id: otel_trace_id
        event_source: event_source
        error:
          schema: schema
          raw: raw
          detail: detail
          type: type
          traceback: traceback
        duration_ms: 0
        completed_at: 2000-01-23T04:56:07.000+00:00
        event_id: event_id
        event_type: event_type
        received_at: 2000-01-23T04:56:07.000+00:00
        namespace: namespace
        schema_output:
          key: ""
        dry_run: true
        id: id
        actions:
        - name: name
          dry_run: true
          type: type
          error: error
          status: status
        - name: name
          dry_run: true
          type: type
          error: error
          status: status
        status: Processing
      properties:
        id:
          type: string
        cluster:
          type: string
        handler:
          type: string
        namespace:
          type: string
        event_id:
          type: string
        event_source:
          type: string
        event_type:
          type: string
        event_raw:
          additionalProperties: true
          type: object
        status:
          $ref: "#/components/schemas/TaskStatus"
        dry_run:
          type: boolean
        schema_output:
          additionalProperties: true
          nullable: true
          type: object
        actions:
          items:
            $ref: "#/components/schemas/ActionResult"
          nullable: true
          type: array
        error:
          $ref: "#/components/schemas/TaskError"
        retry_of:
          nullable: true
          type: string
        otel_trace_id:
          nullable: true
          type: string
        received_at:
          format: date-time
          type: string
        completed_at:
          format: date-time
          nullable: true
          type: string
        duration_ms:
          nullable: true
          type: integer
      required:
      - cluster
      - dry_run
      - event_id
      - event_raw
      - event_source
      - event_type
      - handler
      - id
      - namespace
      - received_at
      - status
      type: object
    CreateTaskRequest:
      example:
        cluster: cluster
        handler: handler
        retry_of: retry_of
        event_raw:
          key: ""
        otel_trace_id: otel_trace_id
        event_source: event_source
        event_id: event_id
        event_type: event_type
        received_at: 2000-01-23T04:56:07.000+00:00
        namespace: namespace
        dry_run: true
        id: id
        status: Processing
      properties:
        id:
          type: string
        cluster:
          type: string
        handler:
          type: string
        namespace:
          type: string
        event_id:
          type: string
        event_source:
          type: string
        event_type:
          type: string
        event_raw:
          additionalProperties: true
          type: object
        status:
          $ref: "#/components/schemas/TaskStatus"
        dry_run:
          type: boolean
        retry_of:
          nullable: true
          type: string
        otel_trace_id:
          nullable: true
          type: string
        received_at:
          format: date-time
          type: string
      required:
      - cluster
      - dry_run
      - event_id
      - event_raw
      - event_source
      - event_type
      - handler
      - id
      - namespace
      - received_at
      - status
      type: object
    UpdateStatusRequest:
      example:
        duration_ms: 0
        completed_at: 2000-01-23T04:56:07.000+00:00
        schema_output:
          key: ""
        error:
          schema: schema
          raw: raw
          detail: detail
          type: type
          traceback: traceback
        actions:
        - name: name
          dry_run: true
          type: type
          error: error
          status: status
        - name: name
          dry_run: true
          type: type
          error: error
          status: status
        status: Processing
      properties:
        status:
          $ref: "#/components/schemas/TaskStatus"
        completed_at:
          format: date-time
          nullable: true
          type: string
        schema_output:
          additionalProperties: true
          nullable: true
          type: object
        actions:
          items:
            $ref: "#/components/schemas/ActionResult"
          nullable: true
          type: array
        error:
          $ref: "#/components/schemas/TaskError"
        duration_ms:
          nullable: true
          type: integer
      required:
      - status
      type: object
    BulkUpdateStatusRequest:
      example:
        ids:
        - ids
        - ids
        status: Processing
      properties:
        ids:
          items:
            type: string
          minItems: 1
          type: array
        status:
          $ref: "#/components/schemas/TaskStatus"
      required:
      - ids
      - status
      type: object
    BulkUpdateStatusResponse:
      example:
        affected_ids:
        - affected_ids
        - affected_ids
      properties:
        affected_ids:
          items:
            type: string
          type: array
      required:
      - affected_ids
      type: object
    TaskList:
      example:
        total: 6
        next_cursor: next_cursor
        tasks:
        - cluster: cluster
          handler: handler
          retry_of: retry_of
          event_raw:
            key: ""
          otel_trace_id: otel_trace_id
          event_source: event_source
          error:
            schema: schema
            raw: raw
            detail: detail
            type: type
            traceback: traceback
          duration_ms: 0
          completed_at: 2000-01-23T04:56:07.000+00:00
          event_id: event_id
          event_type: event_type
          received_at: 2000-01-23T04:56:07.000+00:00
          namespace: namespace
          schema_output:
            key: ""
          dry_run: true
          id: id
          actions:
          - name: name
            dry_run: true
            type: type
            error: error
            status: status
          - name: name
            dry_run: true
            type: type
            error: error
            status: status
          status: Processing
        - cluster: cluster
          handler: handler
          retry_of: retry_of
          event_raw:
            key: ""
          otel_trace_id: otel_trace_id
          event_source: event_source
          error:
            schema: schema
            raw: raw
            detail: detail
            type: type
            traceback: traceback
          duration_ms: 0
          completed_at: 2000-01-23T04:56:07.000+00:00
          event_id: event_id
          event_type: event_type
          received_at: 2000-01-23T04:56:07.000+00:00
          namespace: namespace
          schema_output:
            key: ""
          dry_run: true
          id: id
          actions:
          - name: name
            dry_run: true
            type: type
            error: error
            status: status
          - name: name
            dry_run: true
            type: type
            error: error
            status: status
          status: Processing
      properties:
        tasks:
          items:
            $ref: "#/components/schemas/Task"
          type: array
        total:
          type: integer
        next_cursor:
          nullable: true
          type: string
      required:
      - tasks
      - total
      type: object
    DecisionDistribution:
      example:
        handler: handler
        distribution:
          key: 0
        since: 2000-01-23T04:56:07.000+00:00
      properties:
        handler:
          type: string
        since:
          format: date-time
          type: string
        distribution:
          additionalProperties:
            type: integer
          type: object
      required:
      - distribution
      - handler
      - since
      type: object
    ErrorResponse:
      example:
        error: error
      properties:
        error:
          type: string
      required:
      - error
      type: object