# task-service/internal/interfaces/http/gen/model_bulk_update_status_request.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * KAPE Task Service
 *
 * Persists Task execution records and streams live updates.
 *
 * API version: 1.0.0
 */

package gen




type BulkUpdateStatusRequest struct {

	Ids []string `json:"ids"`

	Status TaskStatus `json:"status"`
}

// AssertBulkUpdateStatusRequestRequired checks if the required fields are not zero-ed
func AssertBulkUpdateStatusRequestRequired(obj BulkUpdateStatusRequest) error {
	elements := map[string]interface{}{
		"ids": obj.Ids,
		"status": obj.Status,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertBulkUpdateStatusRequestConstraints checks if the values respects the defined constraints
func AssertBulkUpdateStatusRequestConstraints(obj BulkUpdateStatusRequest) error {
	return nil
}